+++ www/config.php
@@ -11,7 +11,7 @@ define('DIR_WEB', get_web_dir());
 /** @cont string Директория ядра (движка) относительно домена. Слеш в конце обязателен! */
 define('DIR_WEB_ENGINE', DIR_WEB.'Boolive/');
 /** @cont string Директория проекта относительно домена. Слеш в конце обязателен! */
-define('DIR_WEB_PROJECT', DIR_WEB.'Site/');
+define('DIR_WEB_PROJECT', DIR_WEB);
 
 /** @cont string Директория сайта на сервере. Слеш в конце обязателен! */
 define('DIR_SERVER', DOCUMENT_ROOT.DIR_WEB);